+++ packages/api-models/src/detection/isData.ts
@@ -2,4 +2,3 @@ import { isObject, isPositiveInteger, ValidationFaultCollector } from "@phylopic
 import { Data } from "../types/Data"
 export const isData = (x: unknown, faultCollector?: ValidationFaultCollector): x is Data =>
     isObject(x, faultCollector) && isPositiveInteger((x as Data).build, faultCollector?.sub("build"))
-export default isData