+++ packages/source-client/src/implementations/pg/constants/EXTERNAL_FIELDS.ts
@@ -1,7 +1,7 @@
 import { External } from "@phylopic/source-models"
 import { Authority, Namespace, ObjectID } from "@phylopic/utils"
 import { EditField } from "../fields/EditField"
-const EXTERNAL_FIELDS: ReadonlyArray<
+export const EXTERNAL_FIELDS: ReadonlyArray<
     EditField<External & { authority: Authority; namespace: Namespace; objectID: ObjectID }>
 > = [
     { column: "authority", insertable: true, property: "authority", type: "character varying", updateable: false },
@@ -10,4 +10,3 @@ const EXTERNAL_FIELDS: ReadonlyArray<
     { column: "object_id", insertable: true, property: "objectID", type: "character varying", updateable: false },
     { column: "title", insertable: true, property: "title", type: "character varying", updateable: true },
 ]
-export default EXTERNAL_FIELDS