+++ packages/source-client/src/implementations/pg/fields/getFields.ts
@@ -1,13 +1,12 @@
 import { ReadField } from "./ReadField"
-const getFields = <T>(fields: ReadonlyArray<(string & keyof T) | ReadField<T>>) => {
+export const getFields = <T>(fields: ReadonlyArray<(string & keyof T) | ReadField<T>>) => {
     return fields
         .map(field =>
             typeof field === "string"
                 ? field
                 : field.column === field.property
-                ? field.column
-                : `${field.column} AS ${JSON.stringify(field.property)}`,
+                  ? field.column
+                  : `${field.column} AS ${JSON.stringify(field.property)}`,
         )
         .join(",")
 }
-export default getFields