+++ packages/source-client/src/implementations/s3/S3Reader.ts
@@ -1,8 +1,8 @@
 import { GetObjectCommand, GetObjectOutput } from "@aws-sdk/client-s3"
 import { Readable } from "../../interfaces/Readable"
 import { S3ClientProvider } from "../../interfaces/S3ClientProvider"
-import exists from "./methods/exists"
-export default class S3Reader<T> implements Readable<T> {
+import { exists } from "./methods/exists"
+export class S3Reader<T> implements Readable<T> {
     constructor(
         protected readonly provider: S3ClientProvider,
         protected readonly bucket: string,