+++ packages/source-client/src/implementations/s3/S3TaggingPatcher.ts
@@ -1,8 +1,8 @@
 import { Tagging } from "@aws-sdk/client-s3"
 import { Patchable } from "../../interfaces"
 import { S3ClientProvider } from "../../interfaces/S3ClientProvider"
-import S3TaggingEditor from "./S3TaggingEditor"
-export default class S3TaggingPatcher<T extends Readonly<Record<string, string | null>>>
+import { S3TaggingEditor } from "./S3TaggingEditor"
+export class S3TaggingPatcher<T extends Readonly<Record<string, string | null>>>
     extends S3TaggingEditor<T>
     implements Patchable<T>
 {