+++ packages/source-client/src/implementations/s3/methods/deleteObject.ts
@@ -1,5 +1,5 @@
 import { DeleteObjectCommand, S3Client } from "@aws-sdk/client-s3"
-const deleteObject = async (client: S3Client, Bucket: string, Key: string) => {
+export const deleteObject = async (client: S3Client, Bucket: string, Key: string) => {
     await client.send(
         new DeleteObjectCommand({
             Bucket,
@@ -7,4 +7,3 @@ const deleteObject = async (client: S3Client, Bucket: string, Key: string) => {
         }),
     )
 }
-export default deleteObject