+++ packages/ui/src/search/hooks/useMatches.ts
@@ -1,7 +1,7 @@
 import { compareStrings } from "@phylopic/utils"
 import { useContext, useMemo } from "react"
-import SearchContext from "../context"
-import getSortIndex from "../utils/getSortIndex"
+import { SearchContext } from "../context"
+import { getSortIndex } from "../utils/getSortIndex"
 export const useMatches = (maxLength = Infinity) => {
     const [state] = useContext(SearchContext) ?? []
     const { externalMatches, internalMatches, text } = state || {}
@@ -14,4 +14,3 @@ export const useMatches = (maxLength = Infinity) => {
         return []
     }, [externalMatches, internalMatches, maxLength, text])
 }
-export default useMatches