+++ packages/ui/src/search/sources/paleobiodb.org/PBDB_URL.ts
@@ -1,2 +1,1 @@
-const PBDB_URL = "https://paleobiodb.org/data1.2"
-export default PBDB_URL
+export const PBDB_URL = "https://paleobiodb.org/data1.2"