+++ packages/utils/src/detection/isObject.ts
@@ -1,5 +1,4 @@
-import invalidate from "../validation/invalidate"
-import type ValidationFaultCollector from "../validation/ValidationFaultCollector"
+import { invalidate } from "../validation/invalidate"
+import { type ValidationFaultCollector } from "../validation/ValidationFaultCollector"
 export const isObject = (x: unknown, faultCollector?: ValidationFaultCollector): x is NonNullable<object> =>
     (typeof x === "object" && x !== null) || invalidate(faultCollector, "Expected an object.")
-export default isObject