+++ packages/utils/src/models/conversion/getImageFileExtension.ts
@@ -1,3 +1,2 @@
 import { ImageMediaType } from "../types"
 export const getImageFileExtension = (type: ImageMediaType) => type.replace("image/", "").replace("+xml", "")
-export default getImageFileExtension