+++ packages/utils/src/models/normalization/normalizeNomina.spec.ts
@@ -1,7 +1,6 @@
-import { expect } from "chai"
-import { describe, it } from "mocha"
+import { describe, expect, it } from "vitest"
 import { Nomen } from "../types/Nomen"
-import normalizeNomina from "./normalizeNomina"
+import { normalizeNomina } from "./normalizeNomina"
 const HOMO_SAPIENS: Nomen = [
     {
         class: "scientific",