+++ packages/utils/src/models/normalization/normalizeUUID.ts
@@ -1,4 +1,3 @@
 import { EMPTY_UUID } from "../constants/EMPTY_UUID"
 import { UUIDish } from "../types/UUIDish"
 export const normalizeUUID = (uuid?: UUIDish) => (typeof uuid === "string" ? uuid.toLowerCase() : EMPTY_UUID)
-export default normalizeUUID