+++ packages/utils/src/types/isFiniteNumber.ts
@@ -1,3 +1,2 @@
-import isNumber from "./isNumber"
+import { isNumber } from "./isNumber"
 export const isFiniteNumber = (x: unknown): x is number => isNumber(x) && isFinite(x)
-export default isFiniteNumber